feat: fix imports tree-shake; micromask/index must not contain funcs, only exports
