+++ script/generate-fixtures.js
@@ -23,7 +23,7 @@ await fs.writeFile(
     'export {parse} from "./lib/parse.js"',
     'export {postprocess} from "./lib/postprocess.js"',
     'export {preprocess} from "./lib/preprocess.js"',
-    'export {micromark}',
+    'export {micromark} from "./micromark.js"',
     'import fs from "node:fs"',
     'import path from "node:path"',
     'import {micromark as core} from "./index.bak.js"',
@@ -31,11 +31,6 @@ await fs.writeFile(
     'const base = path.join("test", "fixtures")',
     'process.on("exit", onexit)',
     'fs.mkdirSync(base, {recursive: true})',
-    'function micromark(value) {',
-    '  console.log(...arguments)',
-    '  if (typeof value === "string") captured.push(value)',
-    '  return core(...arguments)',
-    '}',
     'function onexit() {',
     '  captured',
     '    .sort()',
