+++ tsconfig.json
@@ -20,6 +20,7 @@
     "**/*.min.js",
     "packages/micromark/lib/",
     "packages/micromark/index.js",
+    "packages/micromark/micromark.js",
     "packages/micromark/stream.js",
     "packages/micromark-core-commonmark/lib/",
     "packages/micromark-core-commonmark/index.js",
